backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
google-cloud-firestore==2.14.0
resend==0.7.0
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.26.0
